Make a better effort at detecting the base address
On arm64, dynamically-loaded binaries have their first page (where the
ELF headers are) mapped non-executable. Thus, checking for the execute
bit caused the base address detection algorithm to fail.

Instead, ensure that we're looking at a file-backed mapping. Read in the
first page and check if it's an ELF file. This does not guarantee that
we're looking at the base address of the application itself. However,
this makes the algorithm more robust and work for all general use cases.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>