Make the memory mapping code machine dependent
Because we cannot currently set the pc register to an arbitrary value
right now, the memory mapping code now differs significantly between
amd64 and arm64. Life is easy on amd64: we can just redirect execution
to syscall. Not so on arm64: we must wait for the moment right before
the application calls syscall. We'll then replace the application's
syscall arguments with our own.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>